ETL and Data Pipelines with Shell, Airflow and Kafka/Labs/Build a Streaming ETL Pipeline using Kafka/requirements.txt
kafka-python==2.0.2
mysql-connector-python==8.0.31
